backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cached-property==2.0.1
click==8.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.2
h11==0.16.0
idna==3.10
inflection==0.5.1
mypy_extensions==1.1.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-orm==1.2.10
starlette==0.48.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
